# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Library Manageemnt API",
        "description": "Library Manageemnt API for Swagger integration",
        "contact": {
            "name": "Rajan Bhatta (+88 0162119979, aobak63@gmail.com)",
            "email": "aobak63@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/bookshelves": {
            "get": {
                "tags": [
                    "Bookshelves"
                ],
                "summary": "Get list of bookshelves",
                "operationId": "b52dce403a02ed90145cc77e78d797da",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "ApiTokenAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bookshelves"
                ],
                "summary": "Create a new bookshelf",
                "operationId": "a082e83c0b991baa754573bd4662f5ca",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "location"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "location": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bookshelf created"
                    }
                },
                "security": [
                    {
                        "ApiTokenAuth": []
                    }
                ]
            }
        },
        "/api/bookshelves/{id}": {
            "get": {
                "tags": [
                    "Bookshelves"
                ],
                "summary": "Get a single bookshelf",
                "operationId": "b1d9565a9f1be25b6296d3ed1b68668d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bookshelf details"
                    }
                },
                "security": [
                    {
                        "ApiTokenAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Bookshelves"
                ],
                "summary": "Update a bookshelf",
                "operationId": "3486f0061228040450f972603123e595",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "location": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bookshelf updated"
                    }
                },
                "security": [
                    {
                        "ApiTokenAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Bookshelves"
                ],
                "summary": "Delete a bookshelf",
                "operationId": "45e4faef3651d8d9a79c02c4d7dbb1aa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Bookshelf deleted"
                    }
                },
                "security": [
                    {
                        "ApiTokenAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "ApiTokenAuth": {
                "type": "apiKey",
                "description": "Token based authentication using X-API-TOKEN header",
                "name": "X-API-TOKEN",
                "in": "header"
            }
        }
    },
...
}